% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_correlation.R
\name{zoom_correlation}
\alias{zoom_correlation}
\title{Zoom Plot Based on Perfect Correlation}
\usage{
zoom_correlation(data, x_var, y_var, threshold = 0.99, zoom_factor = 0.05)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{x_var}{Name of the x variable.}

\item{y_var}{Name of the y variable.}

\item{threshold}{Correlation threshold (default is 0.99).}

\item{zoom_factor}{Scaling factor for the zoom effect (default is 0.05).}
}
\value{
A ggplot object with zoom applied if correlation threshold is met.
}
\description{
Zoom Plot Based on Perfect Correlation
}
